Update state_client.py
Adjusted "state_client"s main thread, to be as a function, called "program", so by importing it and running "program" function in a non-deamon thread, will allowing the file which is state_client is imported to. to read and write state_client variables, while its main thread is running.

This is to allow the creating of vision-based smart controller script, which reads the gimbal angles, TGP mods, and control mods and then outputs the controller movement.